requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pyyaml==6.0.1
python-dotenv==1.0.0
logging==0.4.9.6